[Refactor] EncryptionService를 EncryptionComponent로 변경 #28
